build: Rename types to core_types to avoid stdlib shadow
This file could potentially shadow python's own types module,
so rename it to core_types.
